refactor Book folder architecture
