+++ package.json
@@ -29,7 +29,7 @@
     "react-lifecycles-compat": "^3.0.4",
     "react-markdown": "^8.0.0",
     "react-router-dom": "^6.2.2",
-    "react-scripts": "^5.0.0",
+    "react-scripts": "^5.0.1",
     "react-style-proptype": "^3.2.2",
     "react-syntax-highlighter": "^15.5.0",
     "real-turtle": "^1.0.21",
